Log all errors
Currently we do not log error in the last case. This case cause an error
to go unnoticed in case there is a retry middleware and the error
disappears.

We had cases when 2 requests were made but only one of them was
logged, so this may help with debugging that case.