Use "private" adapter options in tesla env
The `adapter` keyword in the options conflicts with an internal keyword
in Tesla: https://github.com/elixir-tesla/tesla/blob/9a4b3ad7733e500d5384cdbb54cbbbef85cbecfd/lib/tesla/adapter.ex#LL79C44-L79C44

Our adapter is not in the format that is expected and the keywords
cannot be merged.

Revert back to using "private" options like before and do not pass
everything else that we have in options.