Move telemetry plugs behind retry plug
Currently, a Telemetry event is only published _after_ retries have
completed (successfully or not), while StatsD events, for example, are
published after every attempt.

We are replacing `Tesla.StatsD` usage with a Telemetry-based solution in
a service, and for this migration to be backwards-compatible, the
Telemetry events also need to be published for all attempts.

I expect this to be the preferable behavior in most cases: if a request
needs to be retried, I'd like to see all the requests that were made,
how long they took, what their status codes were, etc. Not a single
event with a long duration.

There's no need to move the OpenTelemetry plug for my immediate goals,
but I expect the above to apply to OpenTelemetry spans as well, so I
moved this as well.