Add support for path params w/ observability
This allows `:telemetry` listeners to use `env.opts[:req_url]` to get
the URL with _unresolved_ path params. This allows the users of this
library who use path params to collect metrics/logs with the unresolved
path params (e.g. `/user/:id` instead of `/user/2`), which makes it
easier to group similar logs/metrics and avoids costs of
high-granularity tags for metrics.

This is roughly based on the [Tesla.Telemetry documentation][1].

Note that this does not revert the changes made in  d341220 ("Use
"private" adapter options in tesla env", 2023-03-17) to "not pass
everything else that we have in options", because previously the
`opts` argument was put into `env.opts` as a whole, but now we're
including the full `env.opts` (not `opts`). This allows other
middlewares that depend on `env.opts` (such as KeepRequest) to continue
working.

[1]: https://hexdocs.pm/tesla/Tesla.Middleware.Telemetry.html#module-url-event-scoping-with-tesla-middleware-pathparams-and-tesla-middleware-keeprequest