Disable StatsD integration by default
Currently it was enabled by default, but in practice it would still not
work by default. That's because the default "backend" is `ExStatsD`,
which is an optional requirement of this library, which means it is not
included in the integrating application by default. So if one were to
set up this library with default configuration, then they'd get a
"module ExStatsD is not available" error when trying to use it.

This also avoids the integrator having to explicitly set `stats: false`
when they want to use the Telemetry integration instead of the StatsD
one.