Bump major version to 4.0.0
While unlikely to affect any users in practice, the change to disable
`StatsD` by default is still a breaking change. Bumping the major
version to avoid any potential issues with this.

It's unlikely to affect any users because it wouldn't work by default
anyway without explicitly installing the `ExStatsD` library. So in
practice it is likely that the `StatsD` middleware, if used, is already
configured anyway. This is indeed the case for all of our own usages of
the library.